+++ CHANGELOG.md
@@ -2,6 +2,11 @@ Changelog
 ---------
 
 
+8.19.2
+------
+* hotfix: only register water administrations once per protocol
+* hotfix: reverse wheel contingency now controlled through task parameter STIM_REVERSE
+
 8.19.1
 ------
 * hotfix: threading warnings during valve calibration (when used without scale)